requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.16
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-redis==4.12.1
django-reversion==3.0.8
djangorestframework==3.12.1
et-xmlfile==1.0.1
future==0.18.2
hiredis==1.1.0
httplib2==0.9.2
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.3.0
